+++ packages/cdn-icon/src/Component.test.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { render, waitFor } from '@testing-library/react';
 
 import { CDNIcon } from './index';
+import * as useIconModule from './hooks/use-icon';
 
 describe('CDNIcon', () => {
     /*
@@ -14,7 +15,11 @@ describe('CDNIcon', () => {
 
     it('should pass an invalid value to the `name` prop', async () => {
         const { container } = render(<CDNIcon name='fake-fake-fake' />);
-        await waitFor(() => expect(container.querySelector('span')).toHaveTextContent(''));
+        await waitFor(() => {
+            const span = container.querySelector('span');
+            expect(span).toBeInTheDocument();
+            expect(span).toBeEmptyDOMElement();
+        });
     });
 
     it('should use the `color` prop', async () => {
@@ -29,5 +34,53 @@ describe('CDNIcon', () => {
         const { container } = render(<CDNIcon name='name' className={className} />);
 
         expect(container.firstElementChild).toHaveClass('class');
+    });
+
+    it('should render fallback node if loading failed', async () => {
+        jest.spyOn(useIconModule, 'useIcon').mockReturnValue([
+            undefined,
+            useIconModule.LoadingStatus.FAILURE,
+        ]);
+
+        const fallbackText = 'fallback-node';
+        const { container } = render(
+            <CDNIcon name='fake' fallback={<span>{fallbackText}</span>} />,
+        );
+
+        await waitFor(() => {
+            const span = container.querySelector('span');
+            expect(span).toBeInTheDocument();
+            expect(span).toHaveTextContent(fallbackText);
+        });
+    });
+
+    it('should call onError if loading failed', async () => {
+        jest.spyOn(useIconModule, 'useIcon').mockReturnValue([
+            undefined,
+            useIconModule.LoadingStatus.FAILURE,
+        ]);
+
+        const onError = jest.fn();
+        render(<CDNIcon name='fake' onError={onError} />);
+
+        await waitFor(() => {
+            expect(onError).toHaveBeenCalled();
+        });
+    });
+
+    it('should render fallback and call onError if both provided', async () => {
+        jest.spyOn(useIconModule, 'useIcon').mockReturnValue([
+            undefined,
+            useIconModule.LoadingStatus.FAILURE,
+        ]);
+
+        const onError = jest.fn();
+        const fallbackText = 'fallback-node';
+
+        render(<CDNIcon name='fake' onError={onError} fallback={<span>{fallbackText}</span>} />);
+
+        await waitFor(() => {
+            expect(onError).toHaveBeenCalled();
+        });
     });
 });